Allow Bitcoin Core 0.10.3 to be 1 block behind

+++ app/models/node.rb
@@ -182,6 +182,9 @@ class Node < ApplicationRecord
       behind = true
     end
 
+    # Allow 1 block extra for 0.10.3 node:
+    return nil if self.client_type == "core" && self.version < 110000 && self.block.height > node.block.height - 2
+
     # Remove entry if no longer behind
     if lag_entry && !behind
       lag_entry.destroy
